Workflow: the development will be done in the develop branch
The development process will be done in the develop branch.
We will try to keep a clean master branch.

First fonctional test added, but not yet passed.
This functional test will be used as an example for
futur functional tests that should be written.

Breaking changes with description of the change, justification and migration notes
Closes|Refs #