Url for retrieving photo informations.
Addition of a new url of the type:
"/en/admin/photo/photo/informations/1/"

or (inside a template):
{% admin:photo-informations 1 %}

Where 1 is a photo id. This url will be called via an ajax request to
retrieve photo informations.

The next step is probably to not return every attributes of the photo
but only some.

Refs Issue#20