Celery: list of tasks for the photo application.
Redefinition of all the tasks from classes to functions.

- uploadimgtoas3:
The task in charge of uploading the picture to the AS3 bucket.

- changeimgsize:
The task in charge of changing the picture size.
In the next steps, we will change the actual value 128x128, to something
more suitable.

- generatewatermarkedimg:
The task in charge of generating a picture with a watermark.